+++ examples/custom-webpack/sanity-app-esm/src/app/app.component.ts
@@ -1,9 +1,11 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
+  imports: [CommonModule],
 })
 export class AppComponent {
   title = 'sanity-app-esm';