+++ examples/jest/multiple-apps/projects/my-shared-library/src/lib/my-shared-library.service.ts
@@ -4,5 +4,4 @@ import { Injectable } from '@angular/core';
   providedIn: 'root',
 })
 export class MySharedLibraryService {
-  constructor() {}
 }