+++ packages/custom-esbuild/src/application/index.ts
@@ -3,11 +3,10 @@ import { BuilderContext, createBuilder } from '@angular-devkit/architect';
 import { buildApplication } from '@angular/build';
 import { getSystemPath, json, normalize } from '@angular-devkit/core';
 import { defer, switchMap } from 'rxjs';
-import { loadModule } from '@angular-builders/common';
 
 import { loadPlugins } from '../load-plugins';
 import { CustomEsbuildApplicationSchema } from '../custom-esbuild-schema';
-import { IndexHtmlTransform } from '@angular/build/private';
+import { loadIndexHtmlTransformer } from '../load-index-html-transformer';
 
 export function buildCustomEsbuildApplication(
   options: CustomEsbuildApplicationSchema,
@@ -17,18 +16,26 @@ export function buildCustomEsbuildApplication(
   const tsConfig = path.join(workspaceRoot, options.tsConfig);
 
   return defer(async () => {
-    const codePlugins = await loadPlugins(options.plugins, workspaceRoot, tsConfig, context.logger);
+    const codePlugins = await loadPlugins(
+      options.plugins,
+      workspaceRoot,
+      tsConfig,
+      context.logger,
+      options,
+      context.target
+    );
 
-    const indexHtmlTransformer: IndexHtmlTransform = options.indexHtmlTransformer
-      ? await loadModule(
+    const indexHtmlTransformer = options.indexHtmlTransformer
+      ? await loadIndexHtmlTransformer(
           path.join(workspaceRoot, options.indexHtmlTransformer),
           tsConfig,
-          context.logger
+          context.logger,
+          context.target
         )
       : undefined;
 
-    return codePlugins;
-  }).pipe(switchMap(plugins => buildApplication(options, context, plugins)));
+    return { codePlugins, indexHtmlTransformer };
+  }).pipe(switchMap(extensions => buildApplication(options, context, extensions)));
 }
 
 export default createBuilder<json.JsonObject & CustomEsbuildApplicationSchema>(