+++ packages/custom-esbuild/src/dev-server/patch-builder-context.ts
@@ -15,9 +15,9 @@ export function patchBuilderContext(context: BuilderContext, buildTarget: Target
   const originalGetBuilderNameForTarget = context.getBuilderNameForTarget;
 
   // We have to patch `getBuilderNameForTarget` because Angular CLI checks
-  // whether the runnable target is `@angular-devkit/build-angular:application`
+  // whether the runnable target is `@angular/build:application`
   // and then defines the server to run. If the `builderName` (returned by
-  // `context.getBuilderNameForTarget`) is not an `@angular-devkit/build-angular:application`,
+  // `context.getBuilderNameForTarget`) is not an `@angular/build:application`,
   // then it will use the Webpack server for the `dev-server target`. By patching
   // the return value, Angular will use the Vite server for the `dev-server` target.
   context.getBuilderNameForTarget = async target => {